requirements.txt
annotated-types==0.7.0
anyio==4.9.0
black==25.1.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
fastapi==0.115.12
feedparser==6.0.11
google-auth==2.40.1
google-genai==1.14.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.59
langsmith==0.3.42
mypy_extensions==1.1.0
nltk==3.9.1
openai==1.78.0
orjson==3.10.18
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pymongo==4.12.1
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
sgmllib3k==1.0.0
sniffio==1.3.1
starlette==0.46.2
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
zstandard==0.23.0